Move unit test about the scene when DPP is enabled from OapStrategySuite.scala to OapFileSourceStrategySuite because OapStrategySuite.scala will be deleted when we abandoned OAP data format
