+++ oap-cache/oap/src/test/scala/org/apache/spark/sql/execution/datasources/oap/OapStrategySuite.scala
@@ -30,7 +30,6 @@ import org.apache.spark.sql.catalyst.plans.logical.CatalystSerde
 import org.apache.spark.sql.execution.QueryExecution
 import org.apache.spark.sql.execution.datasources.oap.adapter.WholeStageCodeGenAdapter
 import org.apache.spark.sql.execution.datasources.oap.utils.OapUtils
-import org.apache.spark.sql.functions.col
 import org.apache.spark.sql.internal.SQLConf
 import org.apache.spark.sql.internal.oap.OapConf
 import org.apache.spark.sql.test.SQLTestUtils
@@ -290,33 +289,6 @@ class OapStrategySuite
     } finally {
       spark.sqlContext.setConf(OapConf.OAP_ENABLE_EXECUTOR_INDEX_SELECTION.key, "false")
       sql("drop oindex index1 on oap_fix_length_schema_table")
-    }
-  }
-
-  test("simple inner join triggers DPP with mock-up tables") {
-    withSQLConf(SQLConf.DYNAMIC_PARTITION_PRUNING_ENABLED.key -> "true",
-      SQLConf.DYNAMIC_PARTITION_PRUNING_REUSE_BROADCAST_ONLY.key -> "true",
-      SQLConf.EXCHANGE_REUSE_ENABLED.key -> "true") {
-      withTable("df1", "df2") {
-        spark.range(1000)
-          .select(col("id"), col("id").as("k"))
-          .write
-          .partitionBy("k")
-          .format("parquet")
-          .mode("overwrite")
-          .saveAsTable("df1")
-
-        spark.range(100)
-          .select(col("id"), col("id").as("k"))
-          .write
-          .partitionBy("k")
-          .format("parquet")
-          .mode("overwrite")
-          .saveAsTable("df2")
-
-        val df = sql("SELECT df1.id, df2.k FROM df1 JOIN df2 ON df1.k = df2.k AND df2.id < 2")
-        checkAnswer(df, Row(0, 0) :: Row(1, 1) :: Nil)
-      }
     }
   }
